+++ src/rogallo/screens/main.py
@@ -55,6 +55,7 @@ from ..commands import (
     Forward,
     GoHome,
     GoToParent,
+    GoToRoot,
     JumpToCommandLine,
     JumpToDocument,
     JumpToSidebar,
@@ -217,6 +218,7 @@ class Main(EnhancedScreen[None]):
         ToggleView,
         GoHome,
         GoToParent,
+        GoToRoot,
         SetHome,
         SetHomeToCurrentLocation,
         SearchHistory,
@@ -379,6 +381,13 @@ class Main(EnhancedScreen[None]):
                 bool(self._viewer.document.location)
                 and isinstance(self._viewer.document.location, GeminiURI)
                 and self._viewer.document.location.parent
+                != self._viewer.document.location
+            )
+        if action == GoToRoot.action_name():
+            return (
+                bool(self._viewer.document.location)
+                and isinstance(self._viewer.document.location, GeminiURI)
+                and self._viewer.document.location.root
                 != self._viewer.document.location
             )
         return True
@@ -1037,6 +1046,14 @@ class Main(EnhancedScreen[None]):
             and location.parent != location
         ):
             self.post_message(OpenLocation(location.parent))
+
+    def action_go_to_root_command(self) -> None:
+        """Go to the root of the current document's location."""
+        if (
+            isinstance(location := self._viewer.document.location, GeminiURI)
+            and location.root != location
+        ):
+            self.post_message(OpenLocation(location.root))
 
 
 ### main.py ends here